# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: API CAIGE
  description: >
    API do sistema CAIGE. Esta especificacao documenta os endpoints atualmente
    montados no backend Express. A documentacao descreve o contrato observavel
    da API e nao substitui as validacoes, autenticacao e autorizacao reais do
    backend.
  version: api-2
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento. A porta padrao vem de PORT ou 3000.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    NaoAutenticado:
      description: Token ausente, invalido ou usuario nao encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroMessage"
    AcessoNegado:
      description: Usuario sem permissao, papel ou escopo necessario.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroMessage"
    ErroInternoMessage:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroMessage"
  schemas:
    ErroMessage:
      type: object
      properties:
        message:
          type: string
          example: Erro interno no servidor.
...
    UsuarioCriacao:
      type: object
      required:
        - email
        - password
        - nome
      properties:
        email:
          type: string
          format: email
          example: professor@univale.br
        password:
          type: string
          format: password
          example: senha-ficticia
        nome:
          type: string
          example: Professor Exemplo
        papel:
          type: string
          enum:
            - SUPERVISOR
            - PROFESSOR
          default: PROFESSOR
        idCurso:
          type: integer
          nullable: true
          description: Obrigatorio quando papel for PROFESSOR.
          example: 1
...
    CursoRequest:
      type: object
      required:
        - nome
      properties:
        nome:
          type: string
          example: Fisioterapia
        name:
          type: string
          description: Alias aceito pelo backend para nome.
          example: Fisioterapia
        ativo:
          type: boolean
          description: Aceito na atualizacao; tambem pode ser enviado como is_active.
          example: true
        is_active:
          type: boolean
          description: Alias aceito pelo backend para ativo.
          example: true
...
    PacienteRequest:
      type: object
      required:
        - name
        - birth_date
        - cpf
        - cep
        - street
        - number
        - neighborhood
        - city
        - state
        - responsible
        - responsible_phone
        - weight_kg
        - height_cm
        - status
      properties:
        name:
          type: string
          example: Paciente Exemplo
        birth_date:
          type: string
          format: date
          example: "2018-04-10"
        gender:
          type: string
          nullable: true
          example: feminino
        cpf:
          type: string
          example: "000.000.000-00"
        phone:
          type: string
          nullable: true
          description: Telefone residencial no padrao (00) 0000-0000. Ao menos phone ou phone2 deve ser informado.
          example: "(33) 3333-3333"
        phone2:
          type: string
          nullable: true
          description: Celular no padrao (00) 00000-0000. Ao menos phone ou phone2 deve ser informado.
          example: "(33) 99999-9999"
        cep:
          type: string
          example: "35000-000"
        street:
          type: string
          example: Rua Exemplo
        number:
          type: string
          example: "100"
        neighborhood:
          type: string
          example: Centro
        city:
          type: string
          example: Governador Valadares
        state:
          type: string
          example: MG
        responsible:
          type: string
          example: Responsavel Exemplo
        responsible_relationship:
          type: string
          nullable: true
          example: Mae
        responsible_phone:
          type: string
          nullable: true
          example: "(33) 99999-0000"
        observations:
          type: string
          nullable: true
          example: Observacao ficticia.
        health_conditions:
          type: string
          nullable: true
          description: Condições de saúde conhecidas do paciente (null quando avaliado e sem ocorrências)
          example: Hipertensão, Diabetes
        health_conditions_evaluated:
          type: boolean
          description: Confirmação de que as condições de saúde foram avaliadas
          example: true
        has_no_health_conditions:
          type: boolean
          description: Confirmação explícita de ausência de condições de saúde
          example: false
        allergies:
          type: string
          nullable: true
          description: Alergias conhecidas do paciente (null quando avaliado e sem ocorrências)
          example: Penicilina, Latex
        allergies_evaluated:
          type: boolean
          description: Confirmação de que as alergias foram avaliadas
          example: true
        has_no_allergies:
          type: boolean
          description: Confirmação explícita de ausência de alergias conhecidas
          example: false
        medications_in_use:
          type: string
          nullable: true
          description: Medicamentos atualmente em uso (null quando avaliado e sem ocorrências)
          example: Losartana 50mg, Metformina 500mg
        medications_in_use_evaluated:
          type: boolean
          description: Confirmação de que os medicamentos em uso foram avaliados
          example: true
        has_no_medications:
          type: boolean
          description: Confirmação explícita de ausência de medicamentos em uso
          example: false
        clinical_observations:
          type: string
          nullable: true
          description: Observações clínicas relevantes (opcional)
          example: Paciente com histórico de cirurgia cardíaca
        weight_kg:
          type: number
          format: double
          nullable: true
          description: Peso atual do paciente em quilogramas
          example: 72.4
        height_cm:
          type: integer
          description: Altura atual do paciente em centímetros
          example: 165
        status:
          type: string
          example: ativo
...
    AtividadeAtendimentoRequest:
      type: object
      required:
        - idCurso
        - nome
      properties:
        idCurso:
          type: integer
          description: Para PROFESSOR, o backend usa req.user.idCurso e bloqueia tentativa de outro curso.
          example: 1
        nome:
          type: string
          example: Alongamento
        descricao:
          type: string
          nullable: true
          example: Sessao leve de alongamento
        ativo:
          type: boolean
          description: Aceito na atualizacao.
          example: true
...
    FrequenciaCriacao:
      type: object
      required:
        - idAtividade
        - idProfissional
      properties:
        idsPacientes:
          type: array
          description: Lista de identificadores numericos dos pacientes (1 a 50). IDs duplicados sao automaticamente deduplicados pelo backend. Mutuamente exclusivo com idPaciente.
          items:
            type: integer
          minItems: 1
          maxItems: 50
          example: [12, 14, 18]
        idPaciente:
          type: integer
          description: Contrato legado para registro de paciente individual. Mutuamente exclusivo com idsPacientes (enviar ambos simultaneamente resulta em HTTP 400).
          example: 12
        idAtividade:
          type: integer
          description: A frequencia e registrada por atividade; a atividade define o curso.
          example: 10
        idProfissional:
          type: integer
          description: Identificador numerico do professor ativo, nao oculto e pertencente ao mesmo curso da atividade. Obrigatorio tanto para SUPERVISOR quanto para PROFESSOR.
          example: 7
        observacao:
          type: string
          nullable: true
          example: Observacao ficticia compartilhada pelo lote.
        observacoes:
          type: string
          nullable: true
          description: Alias aceito pelo backend para observacao.
...
    QuestionarioRequest:
      type: object
      properties:
        title:
          type: string
          example: Avaliacao inicial
        titulo:
          type: string
          description: Alias aceito pelo backend.
        description:
          type: string
          nullable: true
        descricao:
          type: string
          nullable: true
        idCurso:
          type: integer
          example: 1
        course_id:
          type: integer
          description: Alias aceito pelo backend.
        courseId:
          type: integer
          description: Alias aceito pelo backend.
        questions:
          type: array
          items:
            type: integer
        perguntas:
          type: array
          items:
            type: integer
...
    PeriodoLetivo:
      type: object
      properties:
        id:
          type: integer
          example: 3
        ano:
          type: integer
          example: 2026
        semestre:
          type: integer
          enum: [1, 2]
          example: 2
        dataInicio:
          type: string
          format: date
          example: "2026-08-01"
        dataFim:
          type: string
          format: date
          example: "2026-12-20"
        status:
          type: string
          enum: [PLANEJADO, ATIVO, ENCERRADO]
          example: ATIVO
        criadoEm:
          type: string
          nullable: true
        atualizadoEm:
          type: string
          nullable: true
    PeriodoLetivoRequest:
      type: object
      required: [ano, semestre, dataInicio, dataFim]
      properties:
        ano:
          type: integer
          minimum: 2000
          maximum: 2100
          example: 2026
        semestre:
          type: integer
          enum: [1, 2]
          example: 2
        dataInicio:
          type: string
          format: date
          example: "2026-08-01"
        dataFim:
          type: string
          format: date
          example: "2026-12-20"
        status:
          type: string
          enum: [PLANEJADO, ATIVO, ENCERRADO]
          default: PLANEJADO
...
security: []
paths:
...
  /api/usuarios:
    get:
      tags: [Usuarios]
      summary: Lista usuarios administrativos
      description: Exige Bearer JWT e papel SUPERVISOR.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista de usuarios nao ocultos.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
    post:
      tags: [Usuarios]
      summary: Cria usuario
      description: Exige Bearer JWT e papel SUPERVISOR.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioCriacao"
      responses:
        "201":
          description: Usuario criado.
        "400":
          description: Dados invalidos, curso inexistente, e-mail duplicado ou obrigatorio ausente.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...
  /api/pacientes:
    get:
      tags: [Pacientes]
      summary: Lista pacientes ativos/nao arquivados
      description: Exige permissao pode_visualizar_paciente.
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Busca textual por termos.
        - name: status
          in: query
          schema:
            type: string
            enum: [ativo, inativo]
          description: Filtro por status do paciente.
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
          description: Número da página para paginação progressiva (iniciando em 1).
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
          description: Deslocamento numérico de registros (calculado automaticamente a partir de page se omitido).
        - name: limit
          in: query
          schema:
            type: integer
          description: Padrao 50, maximo 100.
      responses:
        "200":
          description: Pacientes retornados.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
    post:
      tags: [Pacientes]
      summary: Cria paciente
      description: Exige permissao pode_criar_paciente.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PacienteRequest"
      responses:
        "201":
          description: Paciente criado.
        "400":
          description: Campos obrigatorios ausentes ou telefone invalido.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...
  /api/dados-painel:
    get:
      tags: [Dados do Painel]
      summary: Retorna indicadores e eventos do dashboard
      description: Para PROFESSOR, os eventos recentes são limitados ao curso de `req.user.idCurso`; os indicadores mantêm a semântica vigente. SUPERVISOR recebe visão global.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dados do painel retornados.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          description: PROFESSOR sem curso vinculado.
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...
  /api/frequencia:
    post:
      tags: [Frequencia]
      summary: Registra frequencia em atividade
      description: >
        Exige permissao pode_registrar_frequencia. Suporta registro individual (idPaciente)
        ou em lote (idsPacientes, de 1 a 50 pacientes), com persistencia transacional atomica (tudo ou nada)
        e bulk insert. Enviar idPaciente e idsPacientes simultaneamente resulta em HTTP 400.
        IDs duplicados no mesmo payload sao automaticamente deduplicados.
        Cada paciente deve estar ativo e nao oculto (caso contrario, a operacao inteira sofre rollback).
        O profissional deve ser um professor ativo, nao oculto e pertencer ao mesmo curso da atividade.
        Para usuarios com papel PROFESSOR, tanto a atividade quanto o profissional devem pertencer ao proprio curso do usuario logado (anti-IDOR).
        Cada paciente do lote recebe uma movimentacao individual correspondente.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrequenciaCriacao"
      responses:
        "201":
          description: Frequencia(s) registrada(s) com sucesso. Retorna totalRegistrados, ids, registradoEm e idPaciente quando formato legado.
        "400":
          description: Parametros ausentes/invalidos, ambiguidade de formato (idPaciente + idsPacientes), lote vazio ou superior a 50 pacientes, paciente inativo/oculto/inexistente, atividade inexistente/inativa ou incompatibilidade de curso.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
  /api/arquivo/usuarios:
    get:
      tags: [Arquivo]
      summary: Lista usuarios arquivados/desabilitados
      description: Exige Bearer JWT e papel SUPERVISOR.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Usuarios arquivados retornados.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
  /api/arquivo/pacientes:
    get:
      tags: [Arquivo]
      summary: Lista pacientes arquivados
      description: Exige Bearer JWT e papel SUPERVISOR. Retorna apenas pacientes com status 'arquivado' e oculto = false.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Pacientes arquivados retornados.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...
  /api/questionarios/questionarios:
    get:
      tags: [Questionarios]
      summary: Lista questionarios para gestao
      description: >
        Exige permissao pode_visualizar_questionarios. PROFESSOR recebe somente
        questionarios do curso vinculado ao usuario. SUPERVISOR pode receber
        todos os cursos ou filtrar por idCurso.
      security:
        - BearerAuth: []
      parameters:
        - name: idCurso
          in: query
          schema:
            type: integer
          description: Filtro opcional por curso para SUPERVISOR.
      responses:
        "200":
          description: Questionarios retornados.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
    post:
      tags: [Questionarios]
      summary: Cria questionario
      description: Exige permissao pode_gerenciar_atividades. A identidade do criador vem de req.user. PROFESSOR usa req.user.idCurso e recebe 403 se tentar informar outro curso.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionarioRequest"
      responses:
        "201":
          description: Questionario criado.
        "400":
          description: Dados invalidos.
        "401":
          description: Usuario autenticado nao encontrado ou token invalido.
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...
  /api/cursos:
    get:
      tags: [Cursos]
      summary: Lista cursos disponiveis
      description: PROFESSOR recebe somente o proprio curso ativo; SUPERVISOR recebe lista completa.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cursos retornados.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
    post:
      tags: [Cursos]
      summary: Cria curso
      description: Exige Bearer JWT e papel SUPERVISOR.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CursoRequest"
      responses:
        "201":
          description: Curso criado.
        "400":
          description: Nome ausente ou duplicado.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...
  /api/atividades-atendimento:
    get:
      tags: [Atividades de Atendimento]
      summary: Lista atividades de atendimento
      description: PROFESSOR fica limitado a req.user.idCurso e nao pode ampliar escopo por query. SUPERVISOR pode filtrar por idCurso.
      security:
        - BearerAuth: []
      parameters:
        - name: idCurso
          in: query
          schema:
            type: integer
          description: Filtro por curso. Bloqueado para PROFESSOR quando difere de req.user.idCurso.
      responses:
        "200":
          description: Atividades retornadas.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
    post:
      tags: [Atividades de Atendimento]
      summary: Cria atividade de atendimento
      description: Exige permissao pode_gerenciar_atividades. PROFESSOR usa req.user.idCurso e recebe 403 se tentar informar outro curso.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtividadeAtendimentoRequest"
      responses:
        "201":
          description: Atividade criada.
        "400":
          description: Curso/nome ausente, curso inativo/inexistente ou duplicidade.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...
  /api/periodos-letivos:
    get:
      tags: [Periodos Letivos]
      summary: Lista periodos letivos
      description: Operacao administrativa exclusiva de SUPERVISOR.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Periodos letivos retornados.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeriodoLetivo"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
        "503":
          description: Migration do calendario ainda nao instalada.
    post:
      tags: [Periodos Letivos]
      summary: Cria periodo letivo
      description: Operacao administrativa exclusiva de SUPERVISOR. Periodos nao podem se sobrepor e so pode existir um periodo ATIVO por vez.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeriodoLetivoRequest"
      responses:
        "201":
          description: Periodo letivo criado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodoLetivo"
        "400":
          description: Ano, semestre, datas ou status invalidos.
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "409":
          description: Conflito por sobreposicao de datas ou outro periodo ativo.
        "500":
          $ref: "#/components/responses/ErroInternoMessage"
...